fix NET_Log declaration

+++ Source/net_client.c
@@ -769,7 +769,7 @@ static void NET_CL_ParseGameData(net_packet_t *packet)
 
         if (!NET_ReadFullTiccmd(packet, &cmd, settings.lowres_turn))
         {
-            NET_Log("client: error: failed to read ticcmd %d", i);
+            NET_Log("client: error: failed to read ticcmd %d", (int)i);
             return;
         }
 
@@ -786,7 +786,7 @@ static void NET_CL_ParseGameData(net_packet_t *packet)
 
         recvobj->active = true;
         recvobj->cmd = cmd;
-        NET_Log("client: stored tic %d in receive window", seq + i);
+        NET_Log("client: stored tic %d in receive window", (int)(seq + i));
 
         // If a packet is lost or arrives out of order, we might get
         // the tic in the next packet instead (because of extratic).
